Store quote date filter preference in local storage

+++ src/utilities/useLocalStorage.ts
@@ -4,8 +4,11 @@ import parseDate from 'date-fns/parse';
 import addDays from 'date-fns/addDays';
 import parseISO from 'date-fns/parseISO';
 
+// This is very specific implementation that handles object that are being passed to it specifically beacuse of dates being matched
+// by the key and not by the type.
+
 const stringifyReplacer = (key: string, value: any) => {
-  if (key === 'date') {
+  if (key === 'date' || key === 'startDate' || key === 'endDate') {
     const dateToParse = value ? (typeof value === 'string' ? parseISO(value) : value) : addDays(new Date(), 2);
     return formatDate(dateToParse, 'yyyy-MM-dd');
   } else {
@@ -14,10 +17,10 @@ const stringifyReplacer = (key: string, value: any) => {
 };
 
 const clearLocalStorageAfter = (lastSavedKey: string, minutes: number) => {
+  if (minutes === -1) return;
   const lastSavedKeyValue = localStorage.getItem(lastSavedKey);
   let saved = lastSavedKeyValue ? parseInt(lastSavedKeyValue) : new Date().getTime();
   if (saved && new Date().getTime() - saved > minutes * 60 * 1000) {
-    console.log('clearing local storage');
     localStorage.clear();
   }
 };
@@ -38,7 +41,7 @@ const useLocalStorage = (key: any, initialValue: any, useRawValues: boolean, tim
         return useRawValues
           ? localStorageValue
           : JSON.parse(localStorageValue || 'null', (key, value) => {
-              if (key === 'date') {
+              if (key === 'date' || key === 'startDate' || key === 'endDate') {
                 return value ? parseDate(value, 'yyyy-MM-dd', addDays(new Date(), 2)) : value;
               } else {
                 return value;
